backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Werkzeug==2.3.7
SQLAlchemy==2.0.21
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==23.0.0
